Added CSS, Folder for Reviews and components
