+++ __test__/ratingsreviews.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import {shallow} from 'enzyme';
-import RatingsReviews from '../client/src/components/ratingsreviews';
+import RatingsReviews from '../client/src/components/reviews/ratingsreviews';
 
 test('Text inside Ratingsreviews', () => {
- 
+
   const wrapper = shallow(<RatingsReviews/>);
 
-  expect(wrapper.text()).toEqual('This is the Ratings Reviews Module');
+  expect(wrapper.text()).toEqual('sort will go here<IRT />MORE REVIEWS');
 
 });
