--- client/src/components/ratingsreviews.jsx
@@ -1,29 +1,0 @@
-import React from "react";
-
-
-class RatingsReviews extends React.Component {
-
-  constructor(props) {
-    super(props);
-
-    //default state
-    this.state = {};
-  }
-
-  render () {
-
-    return (
-      <table border='1px'>
-      <tr>
-        <th>
-            <div>This is the Ratings Reviews Module</div>
-       </th>
-      </tr>
-    </table>
-    )
-  }
-}
-
-
-
-export default RatingsReviews;
