+++ client/src/components/reviews/irt.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+function IRT(props) {
+  return(
+    <div>
+      <div>something here using the IRT things</div>
+      <div>something here using the IRT things</div>
+      <div>something here using the IRT things</div>
+      <div>something here using the IRT things</div>
+      <div>something here using the IRT things</div>
+    </div>
+  );
+}
+
+export default IRT;
