+++ client/src/index.jsx
@@ -1,21 +1,20 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import React from 'react';
+import ReactDOM from 'react-dom';
 import Overview from './components/overview.jsx';
 import QuestionAnswer from './components/questionanswer.jsx';
-import RatingsReviews from './components/ratingsreviews.jsx';
+import RatingsReviews from './components/reviews/ratingsreviews.jsx';
 import RelatedItemsModule from './components/relatedItems/relatedItemsModule.jsx';
 import axios from 'axios';
 
-
 class App extends React.Component {
-
   constructor(props) {
     super(props);
 
-    //default state
+    // default state
     this.state = {
       // Tim will need a property for current product id to make a GET request for related items
       // The team will need a product review rating to render on multiple components
+      reviews: {}
     };
   }
 
@@ -30,7 +29,12 @@ class App extends React.Component {
         <br></br>
         <QuestionAnswer/>
         <br></br>
-        <RatingsReviews/>
+        <div className="reviews-container">
+          <div className="reviews-column-container">
+            <div className="reviews-left">Ratings and Product</div>
+            <div className="reviews-right"><RatingsReviews /></div>
+          </div>
+        </div>
       </div>
     );
